Ajout de la page inscription et connexion

+++ app/src/Service/AddressService.php
@@ -35,15 +35,32 @@ class AddressService
         return $address;
     }
 
-    public function update(Address $address): Address
+    public function update(int $id, Address $address): Address
     {
+        $existingAddress = $this->getAddressById($id);
+
+        if ($existingAddress === null) {
+            throw new \Exception("Address with id $id not found");
+        }
+
+        $existingAddress->setName($address->getName());
+        $existingAddress->setCity($address->getCity());
+        $existingAddress->setCode($address->getCode());
+        $existingAddress->setCountry($address->getCountry());
+
         $this->entityManager->flush();
 
-        return $address;
+        return $existingAddress;
     }
 
-    public function delete(Address $address): void
+    public function delete(int $id): void
     {
+        $address = $this->getAddressById($id);
+
+        if ($address === null) {
+            throw new \Exception("Address with id $id not found");
+        }
+
         $this->entityManager->remove($address);
         $this->entityManager->flush();
     }
